tech-debt(idenplane-js): dedup continuous-verification.ts, split react.ts (#1324)
continuous-verification.ts: extract a private authenticatedPost(path, body,
errorContext) helper handling token retrieval, URL construction, Bearer
auth, and response validation once. reportDevicePosture,
recordBehavioralSample, recordBehavioralSamplesBatch, and
reportNetworkContext now each build their request body and delegate to it,
replacing ~140 lines of repeated fetch/error-handling boilerplate with a
single shared implementation. Added a test file (previously none existed)
covering all four methods plus the authenticatedPost error paths.

react.ts: split into provider.tsx (AuthProvider), context.ts (the shared
AuthContext/useAuthContext internal plumbing extracted so it isn't
duplicated across the split), hooks/ (useAuth, useUser, usePermissions,
useRoles), ProtectedRoute.tsx, and deprecated.ts (IdenplaneProvider,
useIdenplane). react.ts itself becomes a barrel re-exporting all of the
above, preserving the package's "idenplane-sdk/react" entry point.

Confirmed the generated dist/react.d.ts's public `export { ... }`
statement is byte-for-byte identical before and after the split — no
public API change. Done after #1299 so useRoles/usePermissions call into
the settled getRolesFromToken-backed role-checking API rather than a
signature that was still moving.

Closes #1301
